Refactor: Update Pokemon ID type to Long and use image URL
This commit changes the data type for Pokemon IDs from `Int` to `Long` across various files. It also updates the `PokedexScreen` to use the `image` field from `PokemonVO` instead of `url` for displaying Pokemon images.

Key changes:
- `AppRoutes.Pokemon`: `id` field changed from `Int` to `Long`.
- `PokemonDetailVo`: `id` field changed from `Int` to `Long`.
- `formatId` function (platform-specific and common): Parameter `id` type changed from `Int` to `Long`.
- `PokedexScreen.kt`:
    - `onClick` lambda in `PokedexRoute`, `PokedexScreen`, and `ItemList` now accepts a `Long` instead of an `Int`.
    - `ItemList` now uses `model.image` for the `Image` composable instead of `model.url`.
    - Preview data `PokemonVO` instances updated to use `image` instead of `url`.
- `PokemonDetailRoute`: `pokemonId` parameter changed from `Int` to `Long`.
- `PokedexViewModel`: Updated to use `PokemonRepository.getAll()` and adds all fetched Pokemon to the list at once.

+++ composeApp/src/androidMain/kotlin/com/example/pokedex/Platform.android.kt
@@ -9,4 +9,4 @@ class AndroidPlatform : Platform {
 
 actual fun getPlatform(): Platform = AndroidPlatform()
 
-actual fun formatId(id: Int) = String.format(Locale.getDefault(), "#%03d", id)
+actual fun formatId(id: Long) = String.format(Locale.getDefault(), "#%03d", id)
