+++ composeApp/src/iosMain/kotlin/com/example/pokedex/Platform.ios.kt
@@ -8,4 +8,4 @@ class IOSPlatform: Platform {
 
 actual fun getPlatform(): Platform = IOSPlatform()
 
-actual fun formatId(id: Int) = "#${id.toString().padStart(3, '0')}"
+actual fun formatId(id: Long) = "#${id.toString().padStart(3, '0')}"
